change index to server
